+++ CHANGELOG.md
@@ -34,6 +34,10 @@ release.
 -->
 ## [Unreleased]
 
+
+
+## [0.7.0]()
+
 ### Added
 - Added a mutual information matcher [#559](https://github.com/USGS-Astrogeology/autocnet/pull/559)
 - Added residual column information to the Points model
@@ -43,6 +47,11 @@ release.
 - `geom_match_simple` defaults to a 3rd order warp for interpolation
 - Speed improvements for place_points_from_cnet dependent on COPY method instead of ORM update
 - License from custom to CC0. Fixes [#607](https://github.com/USGS-Astrogeology/autocnet/issues/607)
+- `place_points_in_overlap` now properly handles ISIS sensor exceptions
+- Complex geometries that failed to find valid, in geometry points now fallback
+  to using a random point distribution method to ensure points are added.
+- Point and Image deletion from the DB now CASCADE to the measures table making
+  modifications via measures easier to manage. 
 
 ### Fixed
 - `update_from_jigsaw` failures due to stale code. Now uses a conntext on the engine to ensure closure